refactor(topology): move topology state related classes to new package
