+++ topology/src/main/java/io/camunda/zeebe/topology/ClusterTopologyManager.java
@@ -10,6 +10,10 @@ package io.camunda.zeebe.topology;
 import io.atomix.primitive.partition.PartitionMetadata;
 import io.camunda.zeebe.scheduler.ConcurrencyControl;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
+import io.camunda.zeebe.topology.state.ClusterChangePlan;
+import io.camunda.zeebe.topology.state.ClusterTopology;
+import io.camunda.zeebe.topology.state.MemberState;
+import io.camunda.zeebe.topology.state.PartitionState;
 import java.io.IOException;
 import java.util.Map;
 import java.util.Map.Entry;
