+++ topology/src/main/java/io/camunda/zeebe/topology/state/PartitionState.java
@@ -5,10 +5,10 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.topology;
+package io.camunda.zeebe.topology.state;
 
-record PartitionState(State state, int priority) {
-  static PartitionState active(final int priority) {
+public record PartitionState(State state, int priority) {
+  public static PartitionState active(final int priority) {
     return new PartitionState(State.ACTIVE, priority);
   }
