+++ topology/src/test/java/io/camunda/zeebe/topology/ClusterTopologyManagerTest.java
@@ -14,6 +14,9 @@ import io.atomix.cluster.MemberId;
 import io.atomix.primitive.partition.PartitionId;
 import io.atomix.primitive.partition.PartitionMetadata;
 import io.camunda.zeebe.scheduler.testing.TestConcurrencyControl;
+import io.camunda.zeebe.topology.state.ClusterTopology;
+import io.camunda.zeebe.topology.state.MemberState;
+import io.camunda.zeebe.topology.state.PartitionState;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
